refactor(job-inputs): extract demo values handling into a separate module
- Created a new demo.ts file to encapsulate the logic for extracting and validating demo input and output values.
- Removed demo values handling from the form.ts file to improve modularity and maintainability.
- Updated index.ts to export the new demo module for easier access.
