+++ web-app/src/lib/job-input/form.ts
@@ -1,11 +1,5 @@
 import { z } from "zod";
 
-import { AgentDemoData } from "@/lib/db";
-import {
-  jobStatusResponseSchema,
-  JobStatusResponseSchemaType,
-} from "@/lib/schemas";
-
 import { makeZodSchemaFromJobInputSchema } from "./form-schema";
 import { JobInputSchemaType } from "./job-input";
 import { JobInputData, JobInputFormIntlPath, ValidJobInputTypes } from "./type";
@@ -62,39 +56,5 @@ const getDefaultValue = (jobInputSchema: JobInputSchemaType) => {
       return null;
     case ValidJobInputTypes.NONE:
       return null;
-  }
-};
-
-export const getDemoValues = (
-  jobInputSchemas: JobInputSchemaType[],
-  demoData: AgentDemoData,
-): { input: JobInputData; output: JobStatusResponseSchemaType } | null => {
-  const { input, output } = demoData;
-  try {
-    const inputParsedResult = jobInputsFormSchema(jobInputSchemas).safeParse(
-      JSON.parse(input),
-    );
-    let inputValue;
-    if (inputParsedResult.success) {
-      inputValue = filterOutNullValues(inputParsedResult.data);
-    }
-
-    const outputParsedResult = jobStatusResponseSchema.safeParse(
-      JSON.parse(output),
-    );
-    let outputValue;
-    if (outputParsedResult.success) {
-      outputValue = outputParsedResult.data;
-    }
-
-    if (inputValue && outputValue) {
-      return { input: inputValue, output: outputValue };
-    } else {
-      console.error("Invalid demo data", { input, output });
-      return null;
-    }
-  } catch (error) {
-    console.error("Invalid JSON for demo data", error);
-    return null;
   }
 };
