+++ web-app/src/lib/job-input/index.ts
@@ -1,3 +1,4 @@
+export * from "./demo";
 export * from "./form";
 export * from "./job-input";
 export * from "./type";
